Expressions evaluate to point positions
Previously, evaluated expressions were used to move the point, and thus
the positions were defined.  However, this does not allow to handle
cases when positions do not exist.  For example, when we want to move to
the beginning of comment but there is no comment on the current line,
the point will not move.  But when we want to move to the beginning of
line and the point is already there, the point will not move as well.
It is not possible to distinguish these cases with the previous
implementation.

So now, instead of finding position after evaluating a moving
expression, the expression should evaluate to the position itself or nil
when there is no such position (as in the above example with
non-existing comment).

* mwim.el (mwim-line-beginning, mwim-code-beginning)
(mwim-line-end, mwim-code-end, mwim-comment-beginning): New functions.
* mwim.el (mwim-first-position, mwim-next-position): Use the value of
evaluated expressions instead of calculating point position.
(mwim-define-command): Adjust to use position functions.